Move all built-in quad transformers to a dedicated class
